Move files for separation of jobs and users
