+++ server/database/jobs/dbcontroller.js
@@ -1,7 +1,7 @@
 'use strict';
 
 let Job = require('./schema.js');
-let controller = require('../controller.js');
+let controller = require('../../controller.js');
 
 module.exports = {
   retrieveAll: (req, res) => {
